Update IDS.py
This was corrected and cleaned by ChatGPT and is the newest version. It adds a basic logging function and outputs it to an ids.log file. It also has added command-line argument parsing using the argparse module to allow you to specify the network interface and ruleset file as command-line arguments.